Refactorization and error handling
Better error handling, stderr is parsed and stacktraces is put in separate files while a short message is written to journal/stdout. Also added timestamps and loglevels (INFO, WARNING, ERROR and FATAL)

In case one thread raises an exception the main thread gets notices using a thread safe queue and terminates the process. This allows systemd to restart the unit.